+++ .github/workflows/cron.yaml
@@ -1,0 +1,35 @@
+name: Update schedules scraper cron job
+
+on:
+  workflow_dispatch:
+  schedule:
+    # midnight America/Los_Angeles:
+    - cron: "0 7 * * *" # PDT (UTC–7)
+
+jobs:
+  run-cron:
+    runs-on: ubuntu-latest
+    timeout-minutes: 360
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v4
+
+      - name: Set up Node.js & pnpm
+        uses: actions/setup-node@v4
+        with:
+          node-version: "22"
+          cache: "pnpm"
+
+      - name: Install dependencies
+        run: pnpm install --frozen-lockfile
+
+      - name: Run cron script
+        run: pnpm run cron
+        env:
+          PORT: ${{ secrets.PORT }}
+          FRONTEND_ORIGIN: ${{ secrets.FRONTEND_ORIGIN }}
+          POSTGRES_PORT: ${{ secrets.POSTGRES_PORT }}
+          POSTGRES_HOST: ${{ secrets.POSTGRES_HOST }}
+          POSTGRES_USER: ${{ secrets.POSTGRES_USER }}
+          POSTGRES_PASSWORD: ${{ secrets.POSTGRES_PASSWORD }}
+          POSTGRES_DB: ${{ secrets.POSTGRES_DB }}
