+++ backend/package.json
@@ -8,10 +8,13 @@
   },
   "scripts": {
     "build": "npx tsc",
-    "start": "cross-env NODE_ENV=staging node -r module-alias/register dist/index.js",
+    "start": "cross-env NODE_ENV=production node -r module-alias/register dist/index.js",
     "prestart": "pnpm run build",
     "dev": "cross-env NODE_ENV=development nodemon --watch src --exec ts-node -r tsconfig-paths/register src/index.ts",
     "dev-staging": "cross-env NODE_ENV=staging nodemon --watch src --exec ts-node -r tsconfig-paths/register src/index.ts",
+    "cron-prod": "cross-env NODE_ENV=production ts-node -r tsconfig-paths/register src/cron.ts",
+    "cron-staging": "cross-env NODE_ENV=staging ts-node -r tsconfig-paths/register src/cron.ts",
+    "cron": "cross-env NODE_ENV=development ts-node -r tsconfig-paths/register src/cron.ts",
     "lint-check": "eslint . && prettier --check .",
     "lint-fix": "eslint . --fix && prettier --write .",
     "prepare": "cd .. && husky",
@@ -51,11 +54,11 @@
     "express-validator": "^7.2.1",
     "http-errors": "^2.0.0",
     "module-alias": "^2.2.3",
-    "node-cron": "^3.0.3",
     "pg": "^8.13.3",
     "puppeteer": "^24.2.0",
     "sequelize": "^6.37.5",
     "sequelize-typescript": "^2.1.6",
+    "serverless-http": "^3.2.0",
     "tsconfig-paths": "^4.2.0",
     "winston": "^3.17.0",
     "winston-daily-rotate-file": "^5.0.0",
