+++ backend/src/api/index.ts
@@ -1,0 +1,4 @@
+import serverless from "serverless-http";
+import app from "@/app";
+
+export default serverless(app);
